Move constructors to types
