+++ fmm/src/fmm.rs
@@ -1,8 +1,8 @@
-//! Implementation of constructors for FMMs as well as the implementation of FmmData and Fmm traits.
+//! Implementation of FmmData and Fmm traits.
 use cauchy::Scalar;
 use itertools::Itertools;
 use num::{Float, ToPrimitive};
-use std::{collections::HashMap, time::Instant};
+use std::time::Instant;
 
 use rlst::{
     algorithms::{linalg::DenseMatrixLinAlgBuilder, traits::svd::Svd},
@@ -21,7 +21,7 @@ use bempp_traits::{
 };
 use bempp_tree::{constants::ROOT, types::single_node::SingleNodeTree};
 
-use crate::types::{ChargeDict, FmmDataAdaptive, FmmDataUniform, FmmDataUniformMatrix, SendPtrMut};
+use crate::types::{FmmDataAdaptive, FmmDataUniform, FmmDataUniformMatrix};
 use crate::{
     pinv::{pinv, SvdScalar},
     types::KiFmmLinear,
@@ -254,608 +254,7 @@ where
     }
 }
 
-/// Implementation of the data structure to store the data for the single node KiFMM.
-impl<T, U, V> FmmDataUniform<KiFmmLinear<SingleNodeTree<V>, T, U, V>, V>
-where
-    T: Kernel<T = V> + ScaleInvariantKernel<T = V>,
-    U: FieldTranslationData<T>,
-    V: Float + Scalar<Real = V> + Default,
-{
-    /// Constructor fo the KiFMM's associated FmmData on a single node.
-    ///
-    /// # Arguments
-    /// `fmm` - A single node KiFMM object.
-    /// `global_charges` - The charge data associated to the point data via unique global indices.
-    pub fn new(
-        fmm: KiFmmLinear<SingleNodeTree<V>, T, U, V>,
-        global_charges: &ChargeDict<V>,
-    ) -> Result<Self, String> {
-        if let Some(keys) = fmm.tree().get_all_keys() {
-            let ncoeffs = fmm.m2l.ncoeffs(fmm.order);
-            let nkeys = keys.len();
-            let leaves = fmm.tree().get_all_leaves().unwrap();
-            let nleaves = leaves.len();
-            let npoints = fmm.tree().get_all_points().unwrap().len();
 
-            let multipoles = vec![V::default(); ncoeffs * nkeys];
-            let locals = vec![V::default(); ncoeffs * nkeys];
-
-            let mut potentials = vec![V::default(); npoints];
-            let mut potentials_send_pointers = vec![SendPtrMut::default(); nleaves];
-
-            let mut charges = vec![V::default(); npoints];
-            let global_indices = vec![0usize; npoints];
-
-            // Lookup leaf coordinates, and assign charges from within the data tree.
-            for (i, g_idx) in fmm
-                .tree()
-                .get_all_global_indices()
-                .unwrap()
-                .iter()
-                .enumerate()
-            {
-                let charge = global_charges.get(g_idx).unwrap();
-                charges[i] = *charge;
-            }
-
-            let mut level_multipoles = vec![Vec::new(); (fmm.tree().get_depth() + 1) as usize];
-            let mut level_locals = vec![Vec::new(); (fmm.tree().get_depth() + 1) as usize];
-            let mut level_index_pointer =
-                vec![HashMap::new(); (fmm.tree().get_depth() + 1) as usize];
-
-            for level in 0..=fmm.tree().get_depth() {
-                let keys = fmm.tree().get_keys(level).unwrap();
-
-                let mut tmp_multipoles = Vec::new();
-                let mut tmp_locals = Vec::new();
-                for (level_idx, key) in keys.iter().enumerate() {
-                    let idx = fmm.tree().key_to_index.get(key).unwrap();
-                    unsafe {
-                        let raw = multipoles.as_ptr().add(ncoeffs * idx) as *mut V;
-                        tmp_multipoles.push(SendPtrMut { raw });
-                        let raw = locals.as_ptr().add(ncoeffs * idx) as *mut V;
-                        tmp_locals.push(SendPtrMut { raw })
-                    }
-                    level_index_pointer[level as usize].insert(*key, level_idx);
-                }
-                level_multipoles[level as usize] = tmp_multipoles;
-                level_locals[level as usize] = tmp_locals;
-            }
-
-            let mut leaf_multipoles = Vec::new();
-            let mut leaf_locals = Vec::new();
-
-            for leaf in fmm.tree.get_all_leaves().unwrap().iter() {
-                let i = fmm.tree.key_to_index.get(leaf).unwrap();
-                unsafe {
-                    let raw = multipoles.as_ptr().add(i * ncoeffs) as *mut V;
-                    leaf_multipoles.push(SendPtrMut { raw });
-
-                    let raw = locals.as_ptr().add(i * ncoeffs) as *mut V;
-                    leaf_locals.push(SendPtrMut { raw });
-                }
-            }
-
-            // Create an index pointer for the charge data
-            let mut index_pointer = 0;
-            let mut charge_index_pointer = vec![(0usize, 0usize); nleaves];
-
-            let mut potential_raw_pointer = potentials.as_mut_ptr();
-
-            let mut scales = vec![V::default(); nleaves * ncoeffs];
-            for (i, leaf) in leaves.iter().enumerate() {
-                // Assign scales
-                let l = i * ncoeffs;
-                let r = l + ncoeffs;
-                scales[l..r]
-                    .copy_from_slice(vec![fmm.kernel.scale(leaf.level()); ncoeffs].as_slice());
-
-                // Assign potential pointers
-                let npoints;
-                if let Some(points) = fmm.tree().get_points(leaf) {
-                    npoints = points.len();
-                } else {
-                    npoints = 0;
-                }
-
-                potentials_send_pointers[i] = SendPtrMut {
-                    raw: potential_raw_pointer,
-                };
-
-                let bounds = (index_pointer, index_pointer + npoints);
-                charge_index_pointer[i] = bounds;
-                index_pointer += npoints;
-                unsafe { potential_raw_pointer = potential_raw_pointer.add(npoints) };
-            }
-
-            let dim = fmm.kernel().space_dimension();
-            let mut upward_surfaces = vec![V::default(); ncoeffs * nkeys * dim];
-            let mut downward_surfaces = vec![V::default(); ncoeffs * nkeys * dim];
-
-            // For each key form both upward and downward check surfaces
-            for (i, key) in keys.iter().enumerate() {
-                let upward_surface =
-                    key.compute_surface(fmm.tree().get_domain(), fmm.order(), fmm.alpha_outer());
-
-                let downward_surface =
-                    key.compute_surface(fmm.tree().get_domain(), fmm.order(), fmm.alpha_inner());
-
-                let l = i * ncoeffs * dim;
-                let r = l + ncoeffs * dim;
-
-                upward_surfaces[l..r].copy_from_slice(&upward_surface);
-                downward_surfaces[l..r].copy_from_slice(&downward_surface);
-            }
-
-            let mut leaf_upward_surfaces = vec![V::default(); ncoeffs * nleaves * dim];
-            let mut leaf_downward_surfaces = vec![V::default(); ncoeffs * nleaves * dim];
-            for (i, leaf) in leaves.iter().enumerate() {
-                let upward_surface =
-                    leaf.compute_surface(fmm.tree().get_domain(), fmm.order(), fmm.alpha_outer());
-
-                let downward_surface =
-                    leaf.compute_surface(fmm.tree().get_domain(), fmm.order(), fmm.alpha_outer());
-
-                let l = i * ncoeffs * dim;
-                let r = l + ncoeffs * dim;
-                leaf_upward_surfaces[l..r].copy_from_slice(&upward_surface);
-                leaf_downward_surfaces[l..r].copy_from_slice(&downward_surface);
-            }
-
-            return Ok(Self {
-                fmm,
-                multipoles,
-                level_multipoles,
-                leaf_multipoles,
-                locals,
-                level_locals,
-                leaf_locals,
-                level_index_pointer,
-                upward_surfaces,
-                downward_surfaces,
-                leaf_upward_surfaces,
-                leaf_downward_surfaces,
-                potentials,
-                potentials_send_pointers,
-                charges,
-                charge_index_pointer,
-                scales,
-                global_indices,
-            });
-        }
-
-        Err("Not a valid tree".to_string())
-    }
-}
-
-/// Implementation of the data structure to store the data for the single node KiFMM.
-impl<T, U, V> FmmDataUniformMatrix<KiFmmLinear<SingleNodeTree<V>, T, U, V>, V>
-where
-    T: Kernel<T = V> + ScaleInvariantKernel<T = V>,
-    U: FieldTranslationData<T>,
-    V: Float + Scalar<Real = V> + Default,
-{
-    /// Constructor fo the KiFMM's associated FmmData on a single node.
-    ///
-    /// # Arguments
-    /// `fmm` - A single node KiFMM object.
-    /// `global_charges` - The charge data associated to the point data via unique global indices.
-    pub fn new(
-        fmm: KiFmmLinear<SingleNodeTree<V>, T, U, V>,
-        global_charges: &[ChargeDict<V>],
-    ) -> Result<Self, String> {
-        if let Some(keys) = fmm.tree().get_all_keys() {
-            if !fmm.tree().adaptive {
-                let ncoeffs = fmm.m2l.ncoeffs(fmm.order);
-                let nkeys = keys.len();
-                let leaves = fmm.tree().get_all_leaves().unwrap();
-                let nleaves = leaves.len();
-                let npoints = fmm.tree().get_all_points().unwrap().len();
-                let ncharge_vectors = global_charges.len();
-
-                let multipoles = vec![V::default(); ncoeffs * nkeys * ncharge_vectors];
-                let locals = vec![V::default(); ncoeffs * nkeys * ncharge_vectors];
-
-                let mut potentials = vec![V::default(); npoints * ncharge_vectors];
-                let mut potentials_send_pointers =
-                    vec![SendPtrMut::default(); nleaves * ncharge_vectors];
-                let mut scales = vec![V::default(); nleaves * ncoeffs];
-
-                let mut charges = vec![V::default(); npoints * ncharge_vectors];
-                let global_indices = vec![0usize; npoints];
-
-                // Lookup leaf coordinates, and assign charges from within the data tree.
-                for (i, charge_dict) in global_charges.iter().enumerate() {
-                    for (j, g_idx) in fmm
-                        .tree()
-                        .get_all_global_indices()
-                        .unwrap()
-                        .iter()
-                        .enumerate()
-                    {
-                        let charge = charge_dict.get(g_idx).unwrap();
-                        charges[i * npoints + j] = *charge;
-                    }
-                }
-
-                // Indexed by level, then by charge vector
-                let mut level_multipoles =
-                    vec![vec![Vec::new(); ncharge_vectors]; (fmm.tree().get_depth() + 1) as usize];
-                let mut level_locals =
-                    vec![vec![Vec::new(); ncharge_vectors]; (fmm.tree().get_depth() + 1) as usize];
-
-                // Indexed by level, gives key level index
-                let mut level_index_pointer =
-                    vec![HashMap::new(); (fmm.tree().get_depth() + 1) as usize];
-
-                for level in 0..=fmm.tree().get_depth() {
-                    let keys = fmm.tree().get_keys(level).unwrap();
-
-                    for charge_vec_idx in 0..ncharge_vectors {
-                        let mut tmp_multipoles = Vec::new();
-                        let mut tmp_locals = Vec::new();
-                        for key in keys.iter() {
-                            let key_idx = fmm.tree().key_to_index.get(key).unwrap();
-                            let charge_vec_displacement = charge_vec_idx * ncoeffs * nkeys;
-                            let expansion_displacement = ncoeffs * key_idx;
-                            unsafe {
-                                let raw = multipoles
-                                    .as_ptr()
-                                    .add(charge_vec_displacement + expansion_displacement)
-                                    as *mut V;
-                                tmp_multipoles.push(SendPtrMut { raw });
-                                let raw = locals
-                                    .as_ptr()
-                                    .add(charge_vec_displacement + expansion_displacement)
-                                    as *mut V;
-                                tmp_locals.push(SendPtrMut { raw })
-                            }
-                        }
-                        level_multipoles[level as usize][charge_vec_idx] = tmp_multipoles;
-                        level_locals[level as usize][charge_vec_idx] = tmp_locals;
-                    }
-                }
-
-                for level in 0..=fmm.tree().get_depth() {
-                    let keys = fmm.tree().get_keys(level).unwrap();
-                    for (level_idx, key) in keys.iter().enumerate() {
-                        level_index_pointer[level as usize].insert(*key, level_idx);
-                    }
-                }
-
-                // Indexed by charge vector
-                let mut leaf_multipoles = vec![Vec::new(); ncharge_vectors];
-                let mut leaf_locals = vec![Vec::new(); ncharge_vectors];
-
-                for charge_vec_idx in 0..ncharge_vectors {
-                    for leaf in fmm.tree().get_all_leaves().unwrap().iter() {
-                        let key_idx = fmm.tree().key_to_index.get(leaf).unwrap();
-                        let charge_vec_displacement = charge_vec_idx * ncoeffs * nkeys;
-                        let expansion_displacement = ncoeffs * key_idx;
-                        unsafe {
-                            let raw = multipoles
-                                .as_ptr()
-                                .add(charge_vec_displacement + expansion_displacement)
-                                as *mut V;
-                            leaf_multipoles[charge_vec_idx].push(SendPtrMut { raw });
-
-                            let raw = locals
-                                .as_ptr()
-                                .add(charge_vec_displacement + expansion_displacement)
-                                as *mut V;
-                            leaf_locals[charge_vec_idx].push(SendPtrMut { raw });
-                        }
-                    }
-                }
-
-                // Create an index pointer for the point data
-                let mut index_pointer = 0;
-                let mut charge_index_pointer = vec![(0usize, 0usize); nleaves];
-
-                // Get raw pointers to the head of each potential vector
-                let mut potential_raw_pointers = vec![potentials.as_mut_ptr(); ncharge_vectors];
-                for i in 0..ncharge_vectors {
-                    potential_raw_pointers[i] =
-                        unsafe { potential_raw_pointers[i].add(i * npoints) };
-                }
-
-                for (i, leaf) in leaves.iter().enumerate() {
-                    // Assign scales
-                    let l = i * ncoeffs;
-                    let r = l + ncoeffs;
-                    scales[l..r]
-                        .copy_from_slice(vec![fmm.kernel.scale(leaf.level()); ncoeffs].as_slice());
-
-                    // Assign potential pointers
-                    let npoints;
-                    if let Some(points) = fmm.tree().get_points(leaf) {
-                        npoints = points.len();
-                    } else {
-                        npoints = 0;
-                    }
-
-                    for j in 0..ncharge_vectors {
-                        potentials_send_pointers[nleaves * j + i] = SendPtrMut {
-                            raw: potential_raw_pointers[j],
-                        };
-                    }
-
-                    // Update charge index pointer
-                    let bounds = (index_pointer, index_pointer + npoints);
-                    charge_index_pointer[i] = bounds;
-                    index_pointer += npoints;
-
-                    // Update raw pointers with the number of points at this leaf
-                    for j in 0..ncharge_vectors {
-                        unsafe {
-                            potential_raw_pointers[j] = potential_raw_pointers[j].add(npoints)
-                        };
-                    }
-                }
-
-                let dim = fmm.kernel().space_dimension();
-                let mut upward_surfaces = vec![V::default(); ncoeffs * nkeys * dim];
-                let mut downward_surfaces = vec![V::default(); ncoeffs * nkeys * dim];
-
-                // For each key form both upward and downward check surfaces
-                for (i, key) in keys.iter().enumerate() {
-                    let upward_surface = key.compute_surface(
-                        fmm.tree().get_domain(),
-                        fmm.order(),
-                        fmm.alpha_outer(),
-                    );
-
-                    let downward_surface = key.compute_surface(
-                        fmm.tree().get_domain(),
-                        fmm.order(),
-                        fmm.alpha_inner(),
-                    );
-
-                    let l = i * ncoeffs * dim;
-                    let r = l + ncoeffs * dim;
-
-                    upward_surfaces[l..r].copy_from_slice(&upward_surface);
-                    downward_surfaces[l..r].copy_from_slice(&downward_surface);
-                }
-
-                let mut leaf_upward_surfaces = vec![V::default(); ncoeffs * nleaves * dim];
-                let mut leaf_downward_surfaces = vec![V::default(); ncoeffs * nleaves * dim];
-                for (i, leaf) in leaves.iter().enumerate() {
-                    let upward_surface = leaf.compute_surface(
-                        fmm.tree().get_domain(),
-                        fmm.order(),
-                        fmm.alpha_outer(),
-                    );
-
-                    let downward_surface = leaf.compute_surface(
-                        fmm.tree().get_domain(),
-                        fmm.order(),
-                        fmm.alpha_outer(),
-                    );
-
-                    let l = i * ncoeffs * dim;
-                    let r = l + ncoeffs * dim;
-                    leaf_upward_surfaces[l..r].copy_from_slice(&upward_surface);
-                    leaf_downward_surfaces[l..r].copy_from_slice(&downward_surface);
-                }
-
-                return Ok(Self {
-                    fmm,
-                    multipoles,
-                    level_multipoles,
-                    leaf_multipoles,
-                    locals,
-                    level_locals,
-                    leaf_locals,
-                    level_index_pointer,
-                    upward_surfaces,
-                    downward_surfaces,
-                    leaf_upward_surfaces,
-                    leaf_downward_surfaces,
-                    potentials,
-                    potentials_send_pointers,
-                    charges,
-                    charge_index_pointer,
-                    scales,
-                    global_indices,
-                });
-            } else {
-                return Err("Not a uniform tree".to_string());
-            }
-        }
-        Err("Not a valid tree, no keys present".to_string())
-    }
-}
-
-// Implementation of the data structure to store the data for the single node KiFMM.
-impl<T, U, V> FmmDataAdaptive<KiFmmLinear<SingleNodeTree<V>, T, U, V>, V>
-where
-    T: Kernel<T = V> + ScaleInvariantKernel<T = V>,
-    U: FieldTranslationData<T>,
-    V: Float + Scalar<Real = V> + Default,
-{
-    /// Constructor fo the KiFMM's associated FmmData on a single node.
-    ///
-    /// # Arguments
-    /// `fmm` - A single node KiFMM object.
-    /// `global_charges` - The charge data associated to the point data via unique global indices.
-    pub fn new(
-        fmm: KiFmmLinear<SingleNodeTree<V>, T, U, V>,
-        global_charges: &ChargeDict<V>,
-    ) -> Result<Self, String> {
-        if let Some(keys) = fmm.tree().get_all_keys() {
-            if fmm.tree().adaptive {
-                let ncoeffs = fmm.m2l.ncoeffs(fmm.order);
-                let nkeys = keys.len();
-                let leaves = fmm.tree().get_all_leaves().unwrap();
-                let nleaves = leaves.len();
-                let npoints = fmm.tree().get_all_points().unwrap().len();
-
-                let multipoles = vec![V::default(); ncoeffs * nkeys];
-                let locals = vec![V::default(); ncoeffs * nkeys];
-
-                let mut potentials = vec![V::default(); npoints];
-                let mut potentials_send_pointers = vec![SendPtrMut::default(); nleaves];
-
-                let mut charges = vec![V::default(); npoints];
-                let global_indices = vec![0usize; npoints];
-
-                // Lookup leaf coordinates, and assign charges from within the data tree.
-                for (i, g_idx) in fmm
-                    .tree()
-                    .get_all_global_indices()
-                    .unwrap()
-                    .iter()
-                    .enumerate()
-                {
-                    let charge = global_charges.get(g_idx).unwrap();
-                    charges[i] = *charge;
-                }
-
-                let mut level_multipoles = vec![Vec::new(); (fmm.tree().get_depth() + 1) as usize];
-                let mut level_locals = vec![Vec::new(); (fmm.tree().get_depth() + 1) as usize];
-                let mut level_index_pointer =
-                    vec![HashMap::new(); (fmm.tree().get_depth() + 1) as usize];
-
-                for level in 0..=fmm.tree().get_depth() {
-                    let keys = fmm.tree().get_keys(level).unwrap();
-
-                    let mut tmp_multipoles = Vec::new();
-                    let mut tmp_locals = Vec::new();
-                    for (level_idx, key) in keys.iter().enumerate() {
-                        let &idx = fmm.tree().key_to_index.get(key).unwrap();
-                        unsafe {
-                            let raw = multipoles.as_ptr().add(idx * ncoeffs) as *mut V;
-                            tmp_multipoles.push(SendPtrMut { raw });
-
-                            let raw = locals.as_ptr().add(idx * ncoeffs) as *mut V;
-                            tmp_locals.push(SendPtrMut { raw })
-                        }
-                        level_index_pointer[level as usize].insert(*key, level_idx);
-                    }
-                    level_multipoles[level as usize] = tmp_multipoles;
-                    level_locals[level as usize] = tmp_locals;
-                }
-
-                let mut leaf_multipoles = Vec::new();
-                let mut leaf_locals = Vec::new();
-
-                for leaf in fmm.tree.get_all_leaves().unwrap().iter() {
-                    let i = fmm.tree.key_to_index.get(leaf).unwrap();
-                    unsafe {
-                        let raw = multipoles.as_ptr().add(i * ncoeffs) as *mut V;
-                        leaf_multipoles.push(SendPtrMut { raw });
-
-                        let raw = locals.as_ptr().add(i * ncoeffs) as *mut V;
-                        leaf_locals.push(SendPtrMut { raw });
-                    }
-                }
-
-                // Create an index pointer for the charge data
-                let mut index_pointer = 0;
-                let mut charge_index_pointer = vec![(0usize, 0usize); nleaves];
-
-                let mut potential_raw_pointer = potentials.as_mut_ptr();
-
-                let mut scales = vec![V::default(); nleaves * ncoeffs];
-                for (i, leaf) in leaves.iter().enumerate() {
-                    // Assign scales
-                    let l = i * ncoeffs;
-                    let r = l + ncoeffs;
-                    scales[l..r]
-                        .copy_from_slice(vec![fmm.kernel.scale(leaf.level()); ncoeffs].as_slice());
-
-                    // Assign potential pointers
-                    let npoints;
-                    if let Some(points) = fmm.tree().get_points(leaf) {
-                        npoints = points.len();
-                    } else {
-                        npoints = 0;
-                    }
-
-                    potentials_send_pointers[i] = SendPtrMut {
-                        raw: potential_raw_pointer,
-                    };
-
-                    let bounds = (index_pointer, index_pointer + npoints);
-                    charge_index_pointer[i] = bounds;
-                    index_pointer += npoints;
-                    unsafe { potential_raw_pointer = potential_raw_pointer.add(npoints) };
-                }
-
-                let dim = fmm.kernel().space_dimension();
-                let mut upward_surfaces = vec![V::default(); ncoeffs * nkeys * dim];
-                let mut downward_surfaces = vec![V::default(); ncoeffs * nkeys * dim];
-
-                // For each key form both upward and downward check surfaces
-                for (i, key) in keys.iter().enumerate() {
-                    let upward_surface = key.compute_surface(
-                        fmm.tree().get_domain(),
-                        fmm.order(),
-                        fmm.alpha_outer(),
-                    );
-
-                    let downward_surface = key.compute_surface(
-                        fmm.tree().get_domain(),
-                        fmm.order(),
-                        fmm.alpha_inner(),
-                    );
-
-                    let l = i * ncoeffs * dim;
-                    let r = l + ncoeffs * dim;
-
-                    upward_surfaces[l..r].copy_from_slice(&upward_surface);
-                    downward_surfaces[l..r].copy_from_slice(&downward_surface);
-                }
-
-                let mut leaf_upward_surfaces = vec![V::default(); ncoeffs * nleaves * dim];
-                let mut leaf_downward_surfaces = vec![V::default(); ncoeffs * nleaves * dim];
-                for (i, leaf) in leaves.iter().enumerate() {
-                    let upward_surface = leaf.compute_surface(
-                        fmm.tree().get_domain(),
-                        fmm.order(),
-                        fmm.alpha_outer(),
-                    );
-
-                    let downward_surface = leaf.compute_surface(
-                        fmm.tree().get_domain(),
-                        fmm.order(),
-                        fmm.alpha_outer(),
-                    );
-
-                    let l = i * ncoeffs * dim;
-                    let r = l + ncoeffs * dim;
-                    leaf_upward_surfaces[l..r].copy_from_slice(&upward_surface);
-                    leaf_downward_surfaces[l..r].copy_from_slice(&downward_surface);
-                }
-
-                return Ok(Self {
-                    fmm,
-                    multipoles,
-                    level_multipoles,
-                    leaf_multipoles,
-                    locals,
-                    level_locals,
-                    leaf_locals,
-                    level_index_pointer,
-                    upward_surfaces,
-                    downward_surfaces,
-                    leaf_upward_surfaces,
-                    leaf_downward_surfaces,
-                    potentials,
-                    potentials_send_pointers,
-                    charges,
-                    charge_index_pointer,
-                    scales,
-                    global_indices,
-                });
-            } else {
-                return Err("Not an adaptive tree".to_string());
-            }
-        }
-        Err("Not a valid tree, no nodes present.".to_string())
-    }
-}
 
 impl<T, U, V, W> KiFmmTrait for KiFmmLinear<T, U, V, W>
 where
